refactor : 도메인 패키지구조 변경
