+++ src/main/java/nextstep/courses/domain/session/Price.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 public class Price {
     private final int price;
