+++ src/main/java/nextstep/courses/domain/session/SessionStatusType.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 public enum SessionStatusType {
     PREPARE, IN_PROGRESS, END
