+++ src/main/java/nextstep/courses/service/CourseService.java
@@ -1,7 +1,7 @@
 package nextstep.courses.service;
 
-import nextstep.courses.domain.Course;
-import nextstep.courses.domain.CourseRepository;
+import nextstep.courses.domain.course.Course;
+import nextstep.courses.domain.course.CourseRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
