+++ src/test/java/nextstep/courses/domain/session/SessionTest.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
