+++ src/test/java/nextstep/courses/domain/session/SessionsTest.java
@@ -1,4 +1,4 @@
-package nextstep.courses.domain;
+package nextstep.courses.domain.session;
 
 import nextstep.courses.RegisterCourseException;
 import org.assertj.core.api.Assertions;
